巡檢計畫編輯完成

+++ MinSheng_MIS/Models/ViewModels/InspectionPlan_ManagementViewModel.cs
@@ -18,12 +18,16 @@ namespace MinSheng_MIS.Models.ViewModels
         #region 巡檢計畫-詳情 DataGrid
         public string InspectationPlan_Read_Data(string IPSN)
         {
+            #region 參數
             var dic_InPlanState = Surfaces.Surface.InspectionPlanState();
             var dic_Shift = Surfaces.Surface.Shift();
             var dic_EMFIState = Surfaces.Surface.EquipmentMaintainFormItemState();
             var dic_EState = Surfaces.Surface.EState();
             var dic_ERFState = Surfaces.Surface.EquipmentReportFormState();
             var dic_RL = Surfaces.Surface.ReportLevel();
+            List<string> AllMaintainItem = new List<string>();
+            List<string> AllRepairItem = new List<string>();
+            #endregion
 
             #region 巡檢計畫
             var IP_SourceTable = db.InspectionPlan.Find(IPSN);
@@ -65,6 +69,7 @@ namespace MinSheng_MIS.Models.ViewModels
                     { "EName", EI_SourceTable.EName },
                     { "MIName", MI_SourceTable.MIName }
                 };
+                AllMaintainItem.Add(EMI_SourceTable.ESN);
                 ME.Add(ME_Row);
             }
             #endregion
@@ -94,6 +99,7 @@ namespace MinSheng_MIS.Models.ViewModels
                     { "Floor", EI_SourceTable.Floor },
                     { "EName", EI_SourceTable.EName }
                 };
+                AllRepairItem.Add(ERF_SourceTable.ESN);
                 RE.Add(RE_Row);
             }
             #endregion
@@ -114,6 +120,22 @@ namespace MinSheng_MIS.Models.ViewModels
                 #region 路徑標題
                 var FI_SourceTable = db.Floor_Info.Find(IPP_item.FSN);
                 var AI_SourceTable = db.AreaInfo.Find(FI_SourceTable.ASN);
+                var EI_FloorEquip = db.EquipmentInfo.Where(x => x.FSN == IPP_item.FSN).Where(x => x.System == "BT");
+
+                #region 樓層設備 
+                JArray Beacon_ja = new JArray();
+
+                foreach (var EI_item in EI_FloorEquip)
+                {
+                    JObject Beacon_jo = new JObject()
+                    {
+                        {"dbId", EI_item.DBID },
+                        {"deviceType", EI_item.EName },
+                        {"deviceName", EI_item.ESN }
+                    };
+                    Beacon_ja.Add(Beacon_jo);
+                }
+                #endregion
 
                 JObject PathSample_jo = new JObject()
                 {
@@ -124,7 +146,7 @@ namespace MinSheng_MIS.Models.ViewModels
                     {"FSN", IPP_item.FSN },
                     {"PathTitle", IPP_item.PathTitle },
                     {"BIMPath", FI_SourceTable.BIMPath },
-                    {"Beacon","" }
+                    {"Beacon",Beacon_ja }
                 }; //路徑資訊
                 #endregion
 
@@ -148,6 +170,28 @@ namespace MinSheng_MIS.Models.ViewModels
                         { "LocationY", DIPP_item.LocationY }
                     };
                     PathSampleRecord_ja.Add(XY_Path);
+                }
+                #endregion
+
+                #region 保養設備
+                foreach (string MainItem in AllMaintainItem)
+                {
+                    MaintainEquipment_ja.Add(MainItem);
+                }
+                #endregion
+
+                #region 維修設備
+                foreach (string RepairItem in AllRepairItem)
+                {
+                    RepairEquipment_ja.Add(RepairItem);
+                }
+                #endregion
+
+                #region 保養+維修
+                var BothItem = AllMaintainItem.Intersect(AllRepairItem);
+                foreach (string Both_item in BothItem)
+                {
+                    BothEquipment_ja.Add(Both_item);
                 }
                 #endregion
 
@@ -195,6 +239,8 @@ namespace MinSheng_MIS.Models.ViewModels
             var dic_EState = Surfaces.Surface.EState();
             var dic_ERFState = Surfaces.Surface.EquipmentReportFormState();
             var dic_RL = Surfaces.Surface.ReportLevel();
+            List<string> AllMaintainItem = new List<string>();
+            List<string> AllRepairItem = new List<string>();
             #endregion
 
             #region 巡檢計畫
@@ -237,6 +283,7 @@ namespace MinSheng_MIS.Models.ViewModels
                     {"EName", EI_SourceTable.EName},
                     {"MIName", MI_SourceTable.MIName}
                 };
+                AllMaintainItem.Add(EMI_SourceTable.ESN);
                 ME_ja.Add(ME_jo);
             }
 
@@ -265,6 +312,7 @@ namespace MinSheng_MIS.Models.ViewModels
                     { "Floor", EI_SourceTable.Floor },
                     { "EName", EI_SourceTable.EName }
                 };
+                AllRepairItem.Add(ERF_SourceTable.ESN);
                 RE_ja.Add(RE_Row);
             }
             #endregion
@@ -285,6 +333,23 @@ namespace MinSheng_MIS.Models.ViewModels
                 #region 路徑標題
                 var FI_SourceTable = db.Floor_Info.Find(IPP_item.FSN);
                 var AI_SourceTable = db.AreaInfo.Find(FI_SourceTable.ASN);
+                var EI_FloorEquip = db.EquipmentInfo.Where(x => x.FSN == IPP_item.FSN).Where(x => x.System == "BT");
+
+                #region 樓層設備 
+                JArray Beacon_ja = new JArray();
+
+                foreach (var EI_item in EI_FloorEquip)
+                {
+                    JObject Beacon_jo = new JObject()
+                    {
+                        {"dbId", EI_item.DBID },
+                        {"deviceType", EI_item.EName },
+                        {"deviceName", EI_item.ESN }
+                    };
+                    Beacon_ja.Add(Beacon_jo);
+                }
+                #endregion
+
 
                 JObject PathSample_jo = new JObject()
                 {
@@ -295,7 +360,7 @@ namespace MinSheng_MIS.Models.ViewModels
                     {"FSN", IPP_item.FSN },
                     {"PathTitle", IPP_item.PathTitle },
                     {"BIMPath", FI_SourceTable.BIMPath },
-                    {"Beacon","" }
+                    {"Beacon",Beacon_ja }
                 }; //路徑資訊
                 #endregion
 
@@ -322,6 +387,28 @@ namespace MinSheng_MIS.Models.ViewModels
                 }
                 #endregion
 
+                #region 保養設備
+                foreach (string MainItem in AllMaintainItem)
+                {
+                    MaintainEquipment_ja.Add(MainItem);
+                }
+                #endregion
+
+                #region 維修設備
+                foreach (string RepairItem in AllRepairItem)
+                {
+                    RepairEquipment_ja.Add(RepairItem);
+                }
+                #endregion
+
+                #region 保養+維修
+                var BothItem = AllMaintainItem.Intersect(AllRepairItem);
+                foreach (string Both_item in BothItem)
+                {
+                    BothEquipment_ja.Add(Both_item);
+                }
+                #endregion
+
                 JObject IPP_Row = new JObject()
                 {
                     { "PathSample", PathSample_jo },
@@ -340,7 +427,7 @@ namespace MinSheng_MIS.Models.ViewModels
                 {"IPSN", IPSN},
                 {"IPName", IP_SourceTable.IPName },
                 {"PlanCreateUserID", IP_SourceTable.PlanCreateUserID },
-                {"PlanDate", IP_SourceTable.PlanDate.ToString("yyyy/MM/dd") },
+                {"PlanDate", IP_SourceTable.PlanDate.ToString("yyyy-MM-dd") },
                 {"PlanState", dic_InPlanState[IP_SourceTable.PlanState] },
                 {"Shift", IP_SourceTable.Shift },
                 {"UserID", InsName_ja },
@@ -359,7 +446,7 @@ namespace MinSheng_MIS.Models.ViewModels
         #endregion
 
         #region 巡檢計畫-編輯 update
-        public string InspectationPlan_Edit_Update(System.Web.Mvc.FormCollection form)
+        public string InspectationPlan_Edit_Update(System.Web.Mvc.FormCollection form, ref int resultCode)
         {
             /*  前端回傳格式
             IPName: IPName,
@@ -375,33 +462,44 @@ namespace MinSheng_MIS.Models.ViewModels
             */
             
             JsonResponseViewModel Jresult = new JsonResponseViewModel();
-           
+            
             try
             {
                 #region 變數宣告
                 string ipsn = form["IPSN"].ToString();
+                JArray NameArray = (JArray)form["UserID"];
+                JArray EMFISN_ja = (JArray)form["MaintainEquipment"];
+                JArray RSN_ja = (JArray)form["RepairEquipment"];
+                JArray InsPlanPaths = (JArray)form["InspectionPlanPaths"];
                 #endregion
 
                 #region 主表更新 InspectionPlan
                 var IP_SourceTable = db.InspectionPlan.Find(ipsn);
+                if (IP_SourceTable == null)
+                {
+                    resultCode = 400;
+                    Jresult.ResponseCode = 400;
+                    Jresult.ResponseMessage = "更新失敗!";
+                    return JsonConvert.SerializeObject(Jresult);
+                }
                 IP_SourceTable.IPName = form["IPName"].ToString();
                 IP_SourceTable.Shift = form["Shift"].ToString();
                 IP_SourceTable.MaintainUserID = form["MaintainUserID"].ToString();
                 IP_SourceTable.RepairUserID = form["RepairUserID"].ToString();
-                JArray ME_ja = (JArray)form["MaintainEquipment"];
-                JArray RE_ja = (JArray)form["RepairEquipment"];
-                IP_SourceTable.MaintainAmount = ME_ja.Count();
-                IP_SourceTable.RepairAmount = RE_ja.Count();
+                IP_SourceTable.MaintainAmount = EMFISN_ja.Count();
+                IP_SourceTable.RepairAmount = RSN_ja.Count();
                 IP_SourceTable.PlanState = "1";
                 db.InspectionPlan.AddOrUpdate(IP_SourceTable);
                 db.SaveChanges();
                 #endregion
 
                 #region 巡檢計畫人員名單 Inspection Plan Member
-                JArray NameArray = (JArray)form["UserID"];
                 var DelInsPlanMember = db.InspectionPlanMember.Where(x => x.IPSN == ipsn);
-                db.InspectionPlanMember.RemoveRange(DelInsPlanMember);//先刪除全部成員
-                db.SaveChanges();
+                if (DelInsPlanMember != null)
+                {
+                    db.InspectionPlanMember.RemoveRange(DelInsPlanMember);//先刪除全部成員
+                    db.SaveChanges();
+                }
 
                 for (int i = 0; i < NameArray.Count(); i++)
                 {
@@ -418,23 +516,213 @@ namespace MinSheng_MIS.Models.ViewModels
                 #endregion
 
                 #region 定期保養項目
+                var IPM_Del = db.InspectionPlanMaintain.Where(x => x.IPSN == ipsn);
+                if (IPM_Del != null)
+                {
+                    db.InspectionPlanMaintain.RemoveRange(IPM_Del);
+                    db.SaveChanges();
+                }
 
+                for (int i = 0; i < EMFISN_ja.Count;i ++)
+                {
+                    string EMFISN = EMFISN_ja[i].ToString();
+
+                    #region 組IPMSN
+                    int IPMSN_Count = i;
+                    string New_IPMSN = "";
+                    IPMSN_Count++;
+                    if (IPMSN_Count < 10)
+                    {
+                        New_IPMSN = ipsn + "_M0" + IPMSN_Count.ToString();
+                    }
+                    else
+                    {
+                        New_IPMSN = ipsn + "_M" + IPMSN_Count.ToString();
+                    }
+                    #endregion
+
+                    InspectionPlanMaintain IPM = new InspectionPlanMaintain()
+                    {
+                        IPMSN = New_IPMSN,
+                        IPSN = ipsn,
+                        EMFISN = EMFISN,
+                        MaintainState = "1"
+                    };
+                    db.InspectionPlanMaintain.Add(IPM);
+
+                    var EMFI_SourceTable = db.EquipmentMaintainFormItem.Find(EMFISN);
+                    EMFI_SourceTable.FormItemState = "2";
+                    db.EquipmentMaintainFormItem.AddOrUpdate(EMFI_SourceTable);
+                    db.SaveChanges();
+                }
                 #endregion
 
                 #region 維修設備
+                var InsPlanRepair_Del = db.InspectionPlanRepair.Where(x => x.IPSN == ipsn);
+                if (InsPlanRepair_Del != null)
+                {
+                    db.InspectionPlanRepair.RemoveRange(InsPlanRepair_Del);
+                    db.SaveChanges();
+                }
 
+                for (int i = 0; i < RSN_ja.Count; i++)
+                {
+                    string RSN = RSN_ja[i].ToString();
+
+                    #region 組IPRSN
+                    int IPRSN_Count = i;
+                    string New_IPRSN = "";
+                    IPRSN_Count++;
+                    if (IPRSN_Count < 10)
+                    {
+                        New_IPRSN = ipsn + "_R0" + IPRSN_Count.ToString();
+                    }
+                    else
+                    {
+                        New_IPRSN = ipsn + "_R" + IPRSN_Count.ToString();
+                    }
+                    #endregion
+
+                    InspectionPlanRepair IPR = new InspectionPlanRepair()
+                    {
+                        IPRSN = New_IPRSN,
+                        IPSN = ipsn,
+                        RSN = RSN,
+                        RepairState = "1"
+                    };
+                    db.InspectionPlanRepair.Add(IPR);
+
+                    var ERF_SourceTable = db.EquipmentReportForm.Find(RSN);
+                    ERF_SourceTable.ReportState = "2";
+                    db.EquipmentReportForm.AddOrUpdate(ERF_SourceTable);
+
+                    db.SaveChanges();
+                }
                 #endregion
 
                 #region 巡檢路線規劃
+                var InsPlanPath_Del = db.InspectionPlanPath.Where(x => x.IPSN == ipsn);
+                var InsPlanFloorPath_Del = db.InspectionPlanFloorPath.Where(x => x.PSN.Contains(ipsn));
+                
+                foreach (var item in InsPlanPath_Del)
+                {
+                    var DrawPath_Del = db.DrawInspectionPlanPath.Where(x => x.PSN == item.PSN);
+                    if (DrawPath_Del != null)
+                    {
+                        db.DrawInspectionPlanPath.RemoveRange(DrawPath_Del);
+                        db.SaveChanges();
+                    }
+                }
+                
+                if (InsPlanPath_Del != null)
+                {
+                    db.InspectionPlanPath.RemoveRange(InsPlanPath_Del);
+                    db.SaveChanges();
+                }
+                if (InsPlanFloorPath_Del != null)
+                {
+                    db.InspectionPlanFloorPath.RemoveRange(InsPlanFloorPath_Del);
+                    db.SaveChanges();
+                }
 
+                int PSN_Count = 1;
+                foreach (var item in InsPlanPaths)
+                {
+                    #region 組PSN
+                    string New_PSN = "";
+                    if (PSN_Count < 10)
+                    {
+                        New_PSN = ipsn + "_0" + PSN_Count.ToString();
+                    }
+                    else
+                    {
+                        New_PSN = ipsn + "_" + PSN_Count.ToString();
+                    }
+                    PSN_Count++;
+                    #endregion
+
+                    #region 巡檢計畫路徑 InspectionPlanPath
+                    JObject PathSample = (JObject)item["PathSample"];
+                    InspectionPlanPath IPP = new InspectionPlanPath() 
+                    { 
+                        PSN = New_PSN,
+                        IPSN = ipsn,
+                        PathTitle = PathSample["PathTitle"].ToString(),
+                        FSN = PathSample["FSN"].ToString()
+                    };
+                    db.InspectionPlanPath.Add(IPP);
+                    db.SaveChanges();
+                    #endregion
+
+                    #region  巡檢計畫單樓層路徑 InspectionPlanFloorPath
+                    JArray Device_ja = (JArray)item["PathSampleOrder"];
+                    int FPSN_Count = 1;
+                    foreach (string ESN in Device_ja)
+                    {
+                        #region 組FPSN
+                        string New_FPSN = "";
+                        if (FPSN_Count < 10)
+                        {
+                            New_FPSN = New_PSN + "_0" + FPSN_Count.ToString();
+                        }
+                        else
+                        {
+                            New_FPSN = New_PSN + "_" + FPSN_Count.ToString();
+                        }
+                        FPSN_Count++;
+                        #endregion
+
+                        InspectionPlanFloorPath IPFP = new InspectionPlanFloorPath()
+                        {
+                            FPSN = New_FPSN,
+                            PSN = New_PSN,
+                            DeviceID = ESN //待調整
+                        };
+                        db.InspectionPlanFloorPath.Add(IPFP);
+                        db.SaveChanges();
+                    }
+                    #endregion
+
+                    #region 巡檢路徑 DrawInspectionPlanPath
+                    JArray XYPath_ja = (JArray)item["PathSampleRecord"];
+                    int DrawPath_Count = 0;
+                    foreach (JObject XYitem in XYPath_ja)
+                    {
+                        #region 組ISN
+                        DrawPath_Count++;
+                        string New_ISN = "";
+                        if (DrawPath_Count < 10)
+                        {
+                            New_ISN = New_PSN + "_0" + DrawPath_Count.ToString();
+                        }
+                        else
+                        {
+                            New_ISN = New_PSN + "_" + DrawPath_Count.ToString();
+                        }
+                        #endregion
+
+                        DrawInspectionPlanPath DIPP = new DrawInspectionPlanPath()
+                        { 
+                            ISN = New_ISN,
+                            PSN = New_PSN,
+                            LocationX = decimal.Parse(XYitem["LocationX"].ToString()),
+                            LocationY = decimal.Parse(XYitem["LocationY"].ToString())
+                        };
+                        db.DrawInspectionPlanPath.Add(DIPP);
+                        db.SaveChanges();
+                    }
+                    #endregion
+                }
                 #endregion
 
+                resultCode = 200;
                 Jresult.ResponseCode = 200;
                 Jresult.ResponseMessage = "更新成功!";
                 return JsonConvert.SerializeObject(Jresult);
             }
             catch (Exception ex)
             {
+                resultCode = 500;
                 Jresult.ResponseCode = 500;
                 Jresult.ResponseMessage = ex.Message;
                 return JsonConvert.SerializeObject(Jresult);
